Fix a bunch of bugs introduced after the refactor
