+++ bigbluebutton-html5/imports/api/users/server/modifiers/addUser.js
@@ -4,7 +4,6 @@ import Logger from '/imports/startup/server/logger';
 import Meetings from '/imports/api/meetings';
 import Users from '/imports/api/users';
 
-import addChat from '/imports/api/chat/server/modifiers/addChat';
 import requestStunTurn from '../methods/requestStunTurn';
 
 export default function addUser(meetingId, user) {
@@ -30,8 +29,8 @@ export default function addUser(meetingId, user) {
       'user.name': user.name,
       'user._sort_name': user.name.trim().toLowerCase(),
       'user.avatarURL': user.avatarURL,
-      'user.set_emoji_time': user.set_emoji_time || new Date(),
-      'user.time_of_joining': new Date(),
+      'user.set_emoji_time': user.set_emoji_time || (new Date()).getTime(),
+      'user.time_of_joining': (new Date()).getTime(),
       'user.emoji_status': user.emoji_status,
       'user.webcam_stream': user.webcam_stream,
       'user.presenter': user.presenter,
@@ -60,7 +59,6 @@ export default function addUser(meetingId, user) {
 
     const { insertedId } = numChanged;
     if (insertedId) {
-      addWelcomeChatMessage(meetingId, userId);
       return Logger.info(`Added user id=${userId} meeting=${meetingId}`);
     }
 
@@ -70,27 +68,4 @@ export default function addUser(meetingId, user) {
   };
 
   return Users.upsert(selector, modifier, cb);
-};
-
-const addWelcomeChatMessage = (meetingId, userId) => {
-  const APP_CONFIG = Meteor.settings.public.app;
-  const CHAT_CONFIG = Meteor.settings.public.chat;
-
-  const Meeting = Meetings.findOne({ meetingId });
-
-  let welcomeMessage = APP_CONFIG.defaultWelcomeMessage
-      .concat(APP_CONFIG.defaultWelcomeMessageFooter)
-      .replace(/%%CONFNAME%%/, Meeting.meetingName);
-
-  const message = {
-    chat_type: CHAT_CONFIG.type_system,
-    message: welcomeMessage,
-    from_color: '0x3399FF',
-    to_userid: userId,
-    from_userid: CHAT_CONFIG.type_system,
-    from_username: '',
-    from_time: new Date(),
-  };
-
-  return addChat(meetingId, message);
 };
